# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Moscow Noivas API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lojas/{lojaId}/equipe:
    parameters:
      - $ref: "#/components/parameters/LojaIdParam"
    get:
      operationId: listEquipe
      tags: [equipe]
      responses:
        "200":
          description: Equipe da loja
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MembroEquipe"
    post:
      operationId: addMembroEquipe
      tags: [equipe]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembroEquipeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembroEquipe"
  /lojas/{lojaId}/equipe/{usuarioId}:
    parameters:
      - $ref: "#/components/parameters/LojaIdParam"
      - name: usuarioId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateMembroEquipe
      tags: [equipe]
      description: >
        `usuarios` é tabela GLOBAL: o vínculo `usuarios_lojas` é conferido ANTES
        de qualquer escrita (E91/B1) e a resposta é 404 quando a pessoa não é
        desta loja. Sem isso, um admin da loja A inativava a dona da loja B.
        Trocar o perfil ou inativar derruba as sessões vivas, na mesma transação.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembroEquipeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembroEquipe"
        "404":
          description: Não há vínculo desta pessoa com esta loja — nada foi escrito
    delete:
      operationId: removeMembroEquipe
      tags: [equipe]
      description: >
        Mesma prova de pertencimento do PATCH (E91/B1) — antes o 204 saía mesmo
        sem remover nada, e as sessões da pessoa caíam de qualquer jeito.
      responses:
        "204":
          description: Removed
        "404":
          description: Não há vínculo desta pessoa com esta loja
  /lojas/{lojaId}/equipe/convites:
    parameters:
      - $ref: "#/components/parameters/LojaIdParam"
    get:
      operationId: listConvitesEquipe
      tags: [equipe]
      description: "Convites PENDENTES (não usados, não expirados). O token sai aqui de propósito: é o admin quem monta e copia o link."
      responses:
        "200":
          description: Convites pendentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Convite"
    post:
      operationId: createConviteEquipe
      tags: [equipe]
      description: >
        Convite por link: o admin manda o link pelo WhatsApp e a própria pessoa
        define a senha — ninguém mais digita a senha do colega. Vale 7 dias,
        uso único. E-mail já membro da loja → 409 CONVIDADO_JA_E_MEMBRO;
        convite pendente do mesmo e-mail → 409 CONVITE_PENDENTE.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConviteInput"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Convite"
  /lojas/{lojaId}/equipe/convites/{conviteId}/reenviar:
    parameters:
      - $ref: "#/components/parameters/LojaIdParam"
      - name: conviteId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: reenviarConviteEquipe
      tags: [equipe]
      description: "Regenera token e validade — o link antigo morre (desejável se vazou no WhatsApp errado)."
      responses:
        "200":
          description: Convite renovado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Convite"
  /lojas/{lojaId}/equipe/convites/{conviteId}:
    parameters:
      - $ref: "#/components/parameters/LojaIdParam"
      - name: conviteId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: cancelarConviteEquipe
      tags: [equipe]
      responses:
        "204":
          description: Cancelado

  # Rotas PÚBLICAS do convite — sem sessão: quem tem o token (256 bits) é a
  # capability. Token via QUERY/corpo, nunca no path: o log de request grava o
  # path e o token não pode cair em log.
  /lojas/{lojaId}/equipe/atividade:
    parameters:
      - $ref: "#/components/parameters/LojaIdParam"
    get:
      operationId: getAtividadeEquipe
      tags: [equipe]
      summary: Log de atividade da equipe — últimos acessos e ações sensíveis
      description: >
        A visão da dona (E18): quando cada membro entrou pela última vez
        (carimbo de login — sessões somem no logout) e quantas ações sensíveis
        fez nos últimos 30 dias, mais o feed recente do audit_log da loja.
        Mesmo gate admin da gestão de equipe.
      responses:
        "200":
          description: Resumo por membro + eventos recentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtividadeEquipe"
...
components:
  parameters:
    LojaIdParam:
      name: lojaId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    MembroEquipe:
      type: object
      required: [usuarioId, lojaId, perfilId, nome, email, perfilNome]
      properties:
        usuarioId: { type: string }
        lojaId: { type: string }
        perfilId: { type: string }
        nome: { type: string }
        email: { type: string }
        perfilNome: { type: string }
        ativo: { type: boolean }
    MembroEquipeInput:
      type: object
      required: [nome, email, senha, perfilId]
      properties:
        nome: { type: string, minLength: 1 }
        email: { type: string, format: email }
        senha: { type: string, minLength: 6 }
        perfilId: { type: string }
    Convite:
      type: object
      required: [id, lojaId, token, nome, email, perfilId, criadoEm, expiraEm]
      properties:
        id: { type: string }
        lojaId: { type: string }
        token: { type: string }
        nome: { type: string }
        email: { type: string }
        perfilId: { type: string }
        perfilNome: { type: ["string", "null"] }
        criadoEm: { type: string, format: date-time }
        expiraEm: { type: string, format: date-time }
    ConviteInput:
      type: object
      required: [nome, email, perfilId]
      properties:
        nome: { type: string, minLength: 1 }
        email: { type: string, format: email }
        perfilId: { type: string }
...
    MembroAtividade:
      type: object
      required: [usuarioId, nome, email, perfilNome, ativo, acoes30d]
      properties:
        usuarioId: { type: string }
        nome: { type: string }
        email: { type: string }
        perfilNome: { type: string }
        ativo: { type: boolean }
        # Último LOGIN do usuário (global, não por loja). Null = nunca entrou.
        ultimoAcesso: { type: ["string", "null"], format: date-time }
        # Linhas do audit_log desta loja nos últimos 30 dias.
        acoes30d: { type: integer }
    AtividadeEquipe:
      type: object
      required: [membros, eventos]
      properties:
        membros:
          type: array
          items:
            $ref: "#/components/schemas/MembroAtividade"
        eventos:
          type: array
          items:
            $ref: "#/components/schemas/AuditoriaItem"
    MembroEquipeUpdate:
      type: object
      properties:
        nome: { type: string }
        perfilId: { type: string }
        ativo: { type: boolean }

    # ── Backup do sistema (E30) ──
...
    AuditoriaItem:
      type: object
      required: [id, acao, entidade, entidadeId, usuarioNome, criadoEm]
      properties:
        id: { type: string }
        # E107/S4 — `acao` é STRING ABERTA, e isso é a correção de uma
        # contradição entre camadas, não relaxamento.
        #
        # O schema do banco diz, desde o E10: "`acao` é texto versionado pela
        # aplicação, não pgEnum: **trilha nova não pode exigir migration para
        # existir**". E `ROTULO_ACAO` do front é frouxo de propósito — cai no
        # código cru — "porque tela velha lendo trilha nova não pode quebrar".
        #
        # O enum FECHADO aqui contradizia as duas coisas, e o preço não era uma
        # linha mal formatada: `ListAuditoriaResponse.parse()` estoura no
        # primeiro registro desconhecido e **a tela inteira da trilha morre**.
        # Aconteceu: uma linha `CONTA_PAGAR_REMOVIDA` no banco de dev derrubou
        # a lista toda, e a suíte E2E foi quem pegou.
        #
        # Os valores conhecidos vivem em `api-server/src/lib/auditoria.ts`
        # (`ACOES_AUDITORIA`), onde o TypeScript cobra o rótulo de cada um.
        acao:
          type: string
          description: >
            Código da ação, versionado pela aplicação (ver ACOES_AUDITORIA no
            api-server). Aberto de propósito: consumidor antigo lendo trilha
            nova mostra o código cru, nunca quebra.
        entidade: { type: string }
        entidadeId: { type: string }
        usuarioId: { type: ["string", "null"] }
        usuarioNome: { type: string, description: "Desnormalizado — sobrevive à saída do autor da equipe" }
        detalhe:
          type: ["object", "null"]
          additionalProperties: true
        criadoEm: { type: string, format: date-time }
...